+++ src/serdes.h
@@ -188,7 +188,7 @@ serdes_schema_t *serdes_schema_get (serdes_t *sd, const char *name, int id,
  */
 SERDES_EXPORT
 serdes_schema_t *serdes_schema_add (serdes_t *sd, const char *name, int id,
-                                    const void *definition, int definition_len,
+                                    const void *definition, size_t definition_len,
                                     char *errstr, int errstr_size);
